Latest YOLOv8 Trainings
Trained YOLOv8 twice, with training 3 being better overall at detecting bottles, bottle necks and bottle openings.

Training 4, on the other hand, is better at detecting the bottles and the bottle necks, but lacking on bottle openings

Please test those in the robolab today, to detect which ones are better with the lightning and conditions of the room.